requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.6.0
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
async-timeout==4.0.2
    # via redis
authlib==1.3.0
    # via -r requirements.in
bcrypt==4.2.0
    # via paramiko
black==23.3.0
    # via -r requirements.in
cachetools==5.3.0
    # via
    #   -r requirements.in
    #   google-auth
certifi==2022.12.7
    # via
    #   -r requirements.in
    #   requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.1.0
    # via
    #   -r requirements.in
    #   requests
click==8.1.3
    # via black
cryptography==42.0.2
    # via
    #   authlib
    #   paramiko
django==4.1.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-redis
django-cors-headers==4.4.0
    # via -r requirements.in
django-redis==5.2.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
google-api-core==2.11.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-api-python-client==2.81.0
    # via -r requirements.in
google-auth==2.16.2
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   gspread
google-auth-httplib2==0.1.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.0.0
    # via
    #   -r requirements.in
    #   gspread
googleapis-common-protos==1.58.0
    # via
    #   -r requirements.in
    #   google-api-core
gspread==5.7.2
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
httplib2==0.21.0
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via
    #   -r requirements.in
    #   requests
isort==5.12.0
    # via -r requirements.in
marshmallow==3.19.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via black
numpy==2.1.1
    # via pandas
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
openpyxl==3.1.5
    # via -r requirements.in
packaging==23.0
    # via
    #   -r requirements.in
    #   black
    #   gunicorn
    #   marshmallow
pandas==2.2.2
    # via -r requirements.in
paramiko==3.4.1
    # via -r requirements.in
pathspec==0.11.1
    # via black
platformdirs==3.5.0
    # via black
protobuf==4.22.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   googleapis-common-protos
psycopg2-binary==2.9.6
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   -r requirements.in
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via
    #   -r requirements.in
    #   google-auth
pycparser==2.21
    # via cffi
pyjwt==2.9.0
    # via -r requirements.in
pynacl==1.5.0
    # via paramiko
pyparsing==3.0.9
    # via
    #   -r requirements.in
    #   httplib2
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2024.1
    # via pandas
ratelimit==2.2.1
    # via -r requirements.in
redis==4.5.4
    # via django-redis
requests==2.28.2
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
rsa==4.9
    # via
    #   -r requirements.in
    #   google-auth
simplejson==3.18.3
    # via -r requirements.in
six==1.16.0
    # via
    #   -r requirements.in
    #   google-auth
    #   google-auth-httplib2
    #   python-dateutil
sqlparse==0.4.3
    # via
    #   -r requirements.in
    #   django
tomli==2.0.1
    # via black
typed-ast==1.5.5
    # via -r requirements.in
tzdata==2024.1
    # via pandas
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   google-api-python-client
urllib3==1.26.15
    # via
    #   -r requirements.in
    #   requests